requirements.txt
# Core Libraries
numpy==2.3.1
pandas==2.2.3
scikit-learn==1.5.2

# Fuzzy Logic System
scikit-fuzzy==0.5.0

# Constraint Satisfaction Problem Solver
python-constraint==1.4.0

# Web Framework
streamlit==1.51.0

# Visualization & Data Serialization
altair==5.3.0
pyarrow==17.0.0

# Streamlit Core Dependencies
blinker==1.9.0
watchdog==6.0.0
requests==2.32.3
typing-extensions==4.12.2
jinja2==3.1.6
protobuf==5.28.3
tenacity==8.3.0
gitpython==3.1.43
pillow==10.4.0
toml==0.10.2
cachetools==5.5.0

# Compatibility Fixes for Python 3.14
idna==3.7
certifi==2025.11.12
urllib3==2.2.3
